Upgrade VS Code to 1.60.0 (#353)
* chore: upgrade vscode to 1.60.0

* chore: upgrade vscode to 1.60.0

* chore: upgrade vscode to 1.60.0

* chore: upgrade vscode to 1.60.0

* chore: upgrade vscode to 1.60.0

* chore: upgrade vscode to 1.60.0

* feat: update extension config

* feat: update vscode-web version

* feat: update test

* feat: update test snapshots

* feat: update test snapshots

* chore: update test config

* chore: update test config

+++ vscode-web-github1s/scripts/sync-code.sh
@@ -9,7 +9,9 @@ echo $APP_ROOT
 function main() {
 	cd ${APP_ROOT}
 	rsync -a src/ lib/vscode/src
-	rsync -a extensions/ lib/vscode/extensions
+	if [ -e extensions ]; then
+		rsync -a extensions/ lib/vscode/extensions
+	fi
 }
 
 main "$@"
